tls: remove tls_legacy, createSecurePair
This is a semver-major change.
Remove SecurePair, createSecure pair from API.
Remove internal Connection class and SecurePair tests.
