+++ src/node_crypto.cc
@@ -13,11 +13,6 @@
 #include "util.h"
 #include "util-inl.h"
 #include "v8.h"
-// CNNIC Hash WhiteList is taken from
-// https://hg.mozilla.org/mozilla-central/raw-file/98820360ab66/security/
-// certverifier/CNNICHashWhitelist.inc
-#include "CNNICHashWhitelist.inc"
-
 #include <errno.h>
 #include <limits.h>  // INT_MAX
 #include <math.h>
@@ -96,34 +91,6 @@ using v8::String;
 using v8::V8;
 using v8::Value;
 
-
-// Subject DER of CNNIC ROOT CA and CNNIC EV ROOT CA are taken from
-// https://hg.mozilla.org/mozilla-central/file/98820360ab66/security/
-// certverifier/NSSCertDBTrustDomain.cpp#l672
-// C = CN, O = CNNIC, CN = CNNIC ROOT
-static const uint8_t CNNIC_ROOT_CA_SUBJECT_DATA[] =
-    "\x30\x32\x31\x0B\x30\x09\x06\x03\x55\x04\x06\x13\x02\x43\x4E\x31\x0E\x30"
-    "\x0C\x06\x03\x55\x04\x0A\x13\x05\x43\x4E\x4E\x49\x43\x31\x13\x30\x11\x06"
-    "\x03\x55\x04\x03\x13\x0A\x43\x4E\x4E\x49\x43\x20\x52\x4F\x4F\x54";
-static const uint8_t* cnnic_p = CNNIC_ROOT_CA_SUBJECT_DATA;
-static X509_NAME* cnnic_name =
-    d2i_X509_NAME(nullptr, &cnnic_p, sizeof(CNNIC_ROOT_CA_SUBJECT_DATA)-1);
-
-// C = CN, O = China Internet Network Information Center, CN = China
-// Internet Network Information Center EV Certificates Root
-static const uint8_t CNNIC_EV_ROOT_CA_SUBJECT_DATA[] =
-    "\x30\x81\x8A\x31\x0B\x30\x09\x06\x03\x55\x04\x06\x13\x02\x43\x4E\x31\x32"
-    "\x30\x30\x06\x03\x55\x04\x0A\x0C\x29\x43\x68\x69\x6E\x61\x20\x49\x6E\x74"
-    "\x65\x72\x6E\x65\x74\x20\x4E\x65\x74\x77\x6F\x72\x6B\x20\x49\x6E\x66\x6F"
-    "\x72\x6D\x61\x74\x69\x6F\x6E\x20\x43\x65\x6E\x74\x65\x72\x31\x47\x30\x45"
-    "\x06\x03\x55\x04\x03\x0C\x3E\x43\x68\x69\x6E\x61\x20\x49\x6E\x74\x65\x72"
-    "\x6E\x65\x74\x20\x4E\x65\x74\x77\x6F\x72\x6B\x20\x49\x6E\x66\x6F\x72\x6D"
-    "\x61\x74\x69\x6F\x6E\x20\x43\x65\x6E\x74\x65\x72\x20\x45\x56\x20\x43\x65"
-    "\x72\x74\x69\x66\x69\x63\x61\x74\x65\x73\x20\x52\x6F\x6F\x74";
-static const uint8_t* cnnic_ev_p = CNNIC_EV_ROOT_CA_SUBJECT_DATA;
-static X509_NAME *cnnic_ev_name =
-    d2i_X509_NAME(nullptr, &cnnic_ev_p,
-                  sizeof(CNNIC_EV_ROOT_CA_SUBJECT_DATA)-1);
 
 static uv_mutex_t* locks;
 
@@ -416,8 +383,8 @@ void SecureContext::Init(const FunctionCallbackInfo<Value>& args) {
                                  SSL_SESS_CACHE_SERVER |
                                  SSL_SESS_CACHE_NO_INTERNAL |
                                  SSL_SESS_CACHE_NO_AUTO_CLEAR);
-  SSL_CTX_sess_set_get_cb(sc->ctx_, SSLWrap<Connection>::GetSessionCallback);
-  SSL_CTX_sess_set_new_cb(sc->ctx_, SSLWrap<Connection>::NewSessionCallback);
+  SSL_CTX_sess_set_get_cb(sc->ctx_, SSLWrap<TLSWrap>::GetSessionCallback);
+  SSL_CTX_sess_set_new_cb(sc->ctx_, SSLWrap<TLSWrap>::NewSessionCallback);
 
   sc->ca_store_ = nullptr;
 }
@@ -2442,682 +2409,6 @@ int SSLWrap<Base>::SetCACerts(SecureContext* sc) {
   SSL_set_client_CA_list(ssl_, list);
   return 1;
 }
-
-
-void Connection::OnClientHelloParseEnd(void* arg) {
-  Connection* conn = static_cast<Connection*>(arg);
-
-  // Write all accumulated data
-  int r = BIO_write(conn->bio_read_,
-                    reinterpret_cast<char*>(conn->hello_data_),
-                    conn->hello_offset_);
-  conn->HandleBIOError(conn->bio_read_, "BIO_write", r);
-  conn->SetShutdownFlags();
-}
-
-
-#ifdef SSL_PRINT_DEBUG
-# define DEBUG_PRINT(...) fprintf (stderr, __VA_ARGS__)
-#else
-# define DEBUG_PRINT(...)
-#endif
-
-
-int Connection::HandleBIOError(BIO *bio, const char* func, int rv) {
-  if (rv >= 0)
-    return rv;
-
-  int retry = BIO_should_retry(bio);
-  (void) retry;  // unused if !defined(SSL_PRINT_DEBUG)
-
-  if (BIO_should_write(bio)) {
-    DEBUG_PRINT("[%p] BIO: %s want write. should retry %d\n",
-                ssl_,
-                func,
-                retry);
-    return 0;
-
-  } else if (BIO_should_read(bio)) {
-    DEBUG_PRINT("[%p] BIO: %s want read. should retry %d\n", ssl_, func, retry);
-    return 0;
-
-  } else {
-    char ssl_error_buf[512];
-    ERR_error_string_n(rv, ssl_error_buf, sizeof(ssl_error_buf));
-
-    HandleScope scope(ssl_env()->isolate());
-    Local<Value> exception =
-        Exception::Error(OneByteString(ssl_env()->isolate(), ssl_error_buf));
-    object()->Set(ssl_env()->error_string(), exception);
-
-    DEBUG_PRINT("[%p] BIO: %s failed: (%d) %s\n",
-                ssl_,
-                func,
-                rv,
-                ssl_error_buf);
-
-    return rv;
-  }
-
-  return 0;
-}
-
-
-int Connection::HandleSSLError(const char* func,
-                               int rv,
-                               ZeroStatus zs,
-                               SyscallStatus ss) {
-  ClearErrorOnReturn clear_error_on_return;
-  (void) &clear_error_on_return;  // Silence unused variable warning.
-
-  if (rv > 0)
-    return rv;
-  if (rv == 0 && zs == kZeroIsNotAnError)
-    return rv;
-
-  int err = SSL_get_error(ssl_, rv);
-
-  if (err == SSL_ERROR_NONE) {
-    return 0;
-
-  } else if (err == SSL_ERROR_WANT_WRITE) {
-    DEBUG_PRINT("[%p] SSL: %s want write\n", ssl_, func);
-    return 0;
-
-  } else if (err == SSL_ERROR_WANT_READ) {
-    DEBUG_PRINT("[%p] SSL: %s want read\n", ssl_, func);
-    return 0;
-
-  } else if (err == SSL_ERROR_WANT_X509_LOOKUP) {
-    DEBUG_PRINT("[%p] SSL: %s want x509 lookup\n", ssl_, func);
-    return 0;
-
-  } else if (err == SSL_ERROR_ZERO_RETURN) {
-    HandleScope scope(ssl_env()->isolate());
-
-    Local<Value> exception =
-        Exception::Error(ssl_env()->zero_return_string());
-    object()->Set(ssl_env()->error_string(), exception);
-    return rv;
-
-  } else if (err == SSL_ERROR_SYSCALL && ss == kIgnoreSyscall) {
-    return 0;
-
-  } else {
-    HandleScope scope(ssl_env()->isolate());
-    BUF_MEM* mem;
-    BIO *bio;
-
-    CHECK(err == SSL_ERROR_SSL || err == SSL_ERROR_SYSCALL);
-
-    // XXX We need to drain the error queue for this thread or else OpenSSL
-    // has the possibility of blocking connections? This problem is not well
-    // understood. And we should be somehow propagating these errors up
-    // into JavaScript. There is no test which demonstrates this problem.
-    // https://github.com/joyent/node/issues/1719
-    bio = BIO_new(BIO_s_mem());
-    if (bio != nullptr) {
-      ERR_print_errors(bio);
-      BIO_get_mem_ptr(bio, &mem);
-      Local<Value> exception = Exception::Error(
-          OneByteString(ssl_env()->isolate(),
-            mem->data,
-            mem->length));
-      object()->Set(ssl_env()->error_string(), exception);
-      BIO_free_all(bio);
-    }
-
-    return rv;
-  }
-
-  return 0;
-}
-
-
-void Connection::ClearError() {
-#ifndef NDEBUG
-  HandleScope scope(ssl_env()->isolate());
-
-  // We should clear the error in JS-land
-  Local<String> error_key = ssl_env()->error_string();
-  Local<Value> error = object()->Get(error_key);
-  CHECK_EQ(error->BooleanValue(), false);
-#endif  // NDEBUG
-}
-
-
-void Connection::SetShutdownFlags() {
-  HandleScope scope(ssl_env()->isolate());
-
-  int flags = SSL_get_shutdown(ssl_);
-
-  if (flags & SSL_SENT_SHUTDOWN) {
-    Local<String> sent_shutdown_key = ssl_env()->sent_shutdown_string();
-    object()->Set(sent_shutdown_key, True(ssl_env()->isolate()));
-  }
-
-  if (flags & SSL_RECEIVED_SHUTDOWN) {
-    Local<String> received_shutdown_key = ssl_env()->received_shutdown_string();
-    object()->Set(received_shutdown_key, True(ssl_env()->isolate()));
-  }
-}
-
-
-void Connection::NewSessionDoneCb() {
-  HandleScope scope(env()->isolate());
-
-  MakeCallback(env()->onnewsessiondone_string(), 0, nullptr);
-}
-
-
-void Connection::Initialize(Environment* env, Local<Object> target) {
-  Local<FunctionTemplate> t = env->NewFunctionTemplate(Connection::New);
-  t->InstanceTemplate()->SetInternalFieldCount(1);
-  t->SetClassName(FIXED_ONE_BYTE_STRING(env->isolate(), "Connection"));
-
-  env->SetProtoMethod(t, "encIn", Connection::EncIn);
-  env->SetProtoMethod(t, "clearOut", Connection::ClearOut);
-  env->SetProtoMethod(t, "clearIn", Connection::ClearIn);
-  env->SetProtoMethod(t, "encOut", Connection::EncOut);
-  env->SetProtoMethod(t, "clearPending", Connection::ClearPending);
-  env->SetProtoMethod(t, "encPending", Connection::EncPending);
-  env->SetProtoMethod(t, "start", Connection::Start);
-  env->SetProtoMethod(t, "close", Connection::Close);
-
-  SSLWrap<Connection>::AddMethods(env, t);
-
-
-#ifdef SSL_CTRL_SET_TLSEXT_SERVERNAME_CB
-  env->SetProtoMethod(t, "getServername", Connection::GetServername);
-  env->SetProtoMethod(t, "setSNICallback",  Connection::SetSNICallback);
-#endif
-
-  target->Set(FIXED_ONE_BYTE_STRING(env->isolate(), "Connection"),
-              t->GetFunction());
-}
-
-
-inline int compar(const void* a, const void* b) {
-  return memcmp(a, b, CNNIC_WHITELIST_HASH_LEN);
-}
-
-
-inline int IsSelfSigned(X509* cert) {
-  return X509_NAME_cmp(X509_get_subject_name(cert),
-                       X509_get_issuer_name(cert)) == 0;
-}
-
-
-inline X509* FindRoot(STACK_OF(X509)* sk) {
-  for (int i = 0; i < sk_X509_num(sk); i++) {
-    X509* cert = sk_X509_value(sk, i);
-    if (IsSelfSigned(cert))
-      return cert;
-  }
-  return nullptr;
-}
-
-
-// Whitelist check for certs issued by CNNIC. See
-// https://blog.mozilla.org/security/2015/04/02
-// /distrusting-new-cnnic-certificates/
-inline CheckResult CheckWhitelistedServerCert(X509_STORE_CTX* ctx) {
-  unsigned char hash[CNNIC_WHITELIST_HASH_LEN];
-  unsigned int hashlen = CNNIC_WHITELIST_HASH_LEN;
-
-  STACK_OF(X509)* chain = X509_STORE_CTX_get1_chain(ctx);
-  CHECK_NE(chain, nullptr);
-  CHECK_GT(sk_X509_num(chain), 0);
-
-  // Take the last cert as root at the first time.
-  X509* root_cert = sk_X509_value(chain, sk_X509_num(chain)-1);
-  X509_NAME* root_name = X509_get_subject_name(root_cert);
-
-  if (!IsSelfSigned(root_cert)) {
-    root_cert = FindRoot(chain);
-    CHECK_NE(root_cert, nullptr);
-    root_name = X509_get_subject_name(root_cert);
-  }
-
-  // When the cert is issued from either CNNNIC ROOT CA or CNNNIC EV
-  // ROOT CA, check a hash of its leaf cert if it is in the whitelist.
-  if (X509_NAME_cmp(root_name, cnnic_name) == 0 ||
-      X509_NAME_cmp(root_name, cnnic_ev_name) == 0) {
-    X509* leaf_cert = sk_X509_value(chain, 0);
-    int ret = X509_digest(leaf_cert, EVP_sha256(), hash,
-                          &hashlen);
-    CHECK(ret);
-
-    void* result = bsearch(hash, WhitelistedCNNICHashes,
-                           ARRAY_SIZE(WhitelistedCNNICHashes),
-                           CNNIC_WHITELIST_HASH_LEN, compar);
-    if (result == nullptr) {
-      sk_X509_pop_free(chain, X509_free);
-      return CHECK_CERT_REVOKED;
-    }
-  }
-
-  sk_X509_pop_free(chain, X509_free);
-  return CHECK_OK;
-}
-
-
-inline int VerifyCallback(int preverify_ok, X509_STORE_CTX* ctx) {
-  // Failure on verification of the cert is handled in
-  // Connection::VerifyError.
-  if (preverify_ok == 0 || X509_STORE_CTX_get_error(ctx) != X509_V_OK)
-    return 1;
-
-  // Server does not need to check the whitelist.
-  SSL* ssl = static_cast<SSL*>(
-      X509_STORE_CTX_get_ex_data(ctx, SSL_get_ex_data_X509_STORE_CTX_idx()));
-
-  if (SSL_is_server(ssl))
-    return 1;
-
-  // Client needs to check if the server cert is listed in the
-  // whitelist when it is issued by the specific rootCAs.
-  CheckResult ret = CheckWhitelistedServerCert(ctx);
-  if (ret == CHECK_CERT_REVOKED)
-    X509_STORE_CTX_set_error(ctx, X509_V_ERR_CERT_REVOKED);
-
-  return ret;
-}
-
-
-#ifdef SSL_CTRL_SET_TLSEXT_SERVERNAME_CB
-int Connection::SelectSNIContextCallback_(SSL *s, int *ad, void* arg) {
-  Connection* conn = static_cast<Connection*>(SSL_get_app_data(s));
-  Environment* env = conn->env();
-  HandleScope scope(env->isolate());
-
-  const char* servername = SSL_get_servername(s, TLSEXT_NAMETYPE_host_name);
-
-  if (servername) {
-    conn->servername_.Reset(env->isolate(),
-                            OneByteString(env->isolate(), servername));
-
-    // Call the SNI callback and use its return value as context
-    if (!conn->sniObject_.IsEmpty()) {
-      conn->sni_context_.Reset();
-
-      Local<Object> sni_obj = PersistentToLocal(env->isolate(),
-                                                conn->sniObject_);
-
-      Local<Value> arg = PersistentToLocal(env->isolate(), conn->servername_);
-      Local<Value> ret = node::MakeCallback(env->isolate(),
-                                            sni_obj,
-                                            env->onselect_string(),
-                                            1,
-                                            &arg);
-
-      // If ret is SecureContext
-      Local<FunctionTemplate> secure_context_constructor_template =
-          env->secure_context_constructor_template();
-      if (secure_context_constructor_template->HasInstance(ret)) {
-        conn->sni_context_.Reset(env->isolate(), ret);
-        SecureContext* sc = Unwrap<SecureContext>(ret.As<Object>());
-        conn->SetSNIContext(sc);
-      } else {
-        return SSL_TLSEXT_ERR_NOACK;
-      }
-    }
-  }
-
-  return SSL_TLSEXT_ERR_OK;
-}
-#endif
-
-void Connection::New(const FunctionCallbackInfo<Value>& args) {
-  Environment* env = Environment::GetCurrent(args);
-
-  if (args.Length() < 1 || !args[0]->IsObject()) {
-    env->ThrowError("First argument must be a tls module SecureContext");
-    return;
-  }
-
-  SecureContext* sc = Unwrap<SecureContext>(args[0]->ToObject(env->isolate()));
-
-  bool is_server = args[1]->BooleanValue();
-
-  SSLWrap<Connection>::Kind kind =
-      is_server ? SSLWrap<Connection>::kServer : SSLWrap<Connection>::kClient;
-  Connection* conn = new Connection(env, args.This(), sc, kind);
-  conn->bio_read_ = NodeBIO::New();
-  conn->bio_write_ = NodeBIO::New();
-
-  SSL_set_app_data(conn->ssl_, conn);
-
-  if (is_server)
-    SSL_set_info_callback(conn->ssl_, SSLInfoCallback);
-
-  InitNPN(sc);
-
-  SSL_set_cert_cb(conn->ssl_, SSLWrap<Connection>::SSLCertCallback, conn);
-
-#ifdef SSL_CTRL_SET_TLSEXT_SERVERNAME_CB
-  if (is_server) {
-    SSL_CTX_set_tlsext_servername_callback(sc->ctx_, SelectSNIContextCallback_);
-  } else if (args[2]->IsString()) {
-    const node::Utf8Value servername(env->isolate(), args[2]);
-    SSL_set_tlsext_host_name(conn->ssl_, *servername);
-  }
-#endif
-
-  SSL_set_bio(conn->ssl_, conn->bio_read_, conn->bio_write_);
-
-#ifdef SSL_MODE_RELEASE_BUFFERS
-  long mode = SSL_get_mode(conn->ssl_);
-  SSL_set_mode(conn->ssl_, mode | SSL_MODE_RELEASE_BUFFERS);
-#endif
-
-
-  int verify_mode;
-  if (is_server) {
-    bool request_cert = args[2]->BooleanValue();
-    if (!request_cert) {
-      // Note reject_unauthorized ignored.
-      verify_mode = SSL_VERIFY_NONE;
-    } else {
-      bool reject_unauthorized = args[3]->BooleanValue();
-      verify_mode = SSL_VERIFY_PEER;
-      if (reject_unauthorized)
-        verify_mode |= SSL_VERIFY_FAIL_IF_NO_PEER_CERT;
-    }
-  } else {
-    // Note request_cert and reject_unauthorized are ignored for clients.
-    verify_mode = SSL_VERIFY_NONE;
-  }
-
-
-  // Always allow a connection. We'll reject in javascript.
-  SSL_set_verify(conn->ssl_, verify_mode, VerifyCallback);
-
-  if (is_server) {
-    SSL_set_accept_state(conn->ssl_);
-  } else {
-    SSL_set_connect_state(conn->ssl_);
-  }
-}
-
-
-void Connection::SSLInfoCallback(const SSL *ssl_, int where, int ret) {
-  if (!(where & (SSL_CB_HANDSHAKE_START | SSL_CB_HANDSHAKE_DONE)))
-    return;
-
-  // Be compatible with older versions of OpenSSL. SSL_get_app_data() wants
-  // a non-const SSL* in OpenSSL <= 0.9.7e.
-  SSL* ssl = const_cast<SSL*>(ssl_);
-  Connection* conn = static_cast<Connection*>(SSL_get_app_data(ssl));
-  Environment* env = conn->env();
-  HandleScope handle_scope(env->isolate());
-  Context::Scope context_scope(env->context());
-
-  if (where & SSL_CB_HANDSHAKE_START) {
-    conn->MakeCallback(env->onhandshakestart_string(), 0, nullptr);
-  }
-
-  if (where & SSL_CB_HANDSHAKE_DONE) {
-    conn->MakeCallback(env->onhandshakedone_string(), 0, nullptr);
-  }
-}
-
-
-void Connection::EncIn(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-  Environment* env = conn->env();
-
-  if (args.Length() < 3) {
-    return env->ThrowTypeError(
-        "Data, offset, and length arguments are mandatory");
-  }
-
-  THROW_AND_RETURN_IF_NOT_BUFFER(args[0], "Data");
-
-  char* buffer_data = Buffer::Data(args[0]);
-  size_t buffer_length = Buffer::Length(args[0]);
-
-  size_t off = args[1]->Int32Value();
-  size_t len = args[2]->Int32Value();
-
-  if (!Buffer::IsWithinBounds(off, len, buffer_length))
-    return env->ThrowRangeError("offset + length > buffer.length");
-
-  int bytes_written;
-  char* data = buffer_data + off;
-
-  if (conn->is_server() && !conn->hello_parser_.IsEnded()) {
-    // Just accumulate data, everything will be pushed to BIO later
-    if (conn->hello_parser_.IsPaused()) {
-      bytes_written = 0;
-    } else {
-      // Copy incoming data to the internal buffer
-      // (which has a size of the biggest possible TLS frame)
-      size_t available = sizeof(conn->hello_data_) - conn->hello_offset_;
-      size_t copied = len < available ? len : available;
-      memcpy(conn->hello_data_ + conn->hello_offset_, data, copied);
-      conn->hello_offset_ += copied;
-
-      conn->hello_parser_.Parse(conn->hello_data_, conn->hello_offset_);
-      bytes_written = copied;
-    }
-  } else {
-    bytes_written = BIO_write(conn->bio_read_, data, len);
-    conn->HandleBIOError(conn->bio_read_, "BIO_write", bytes_written);
-    conn->SetShutdownFlags();
-  }
-
-  args.GetReturnValue().Set(bytes_written);
-}
-
-
-void Connection::ClearOut(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-  Environment* env = conn->env();
-
-  if (args.Length() < 3) {
-    return env->ThrowTypeError(
-        "Data, offset, and length arguments are mandatory");
-  }
-
-  THROW_AND_RETURN_IF_NOT_BUFFER(args[0], "Data");
-
-  char* buffer_data = Buffer::Data(args[0]);
-  size_t buffer_length = Buffer::Length(args[0]);
-
-  size_t off = args[1]->Int32Value();
-  size_t len = args[2]->Int32Value();
-
-  if (!Buffer::IsWithinBounds(off, len, buffer_length))
-    return env->ThrowRangeError("offset + length > buffer.length");
-
-  if (!SSL_is_init_finished(conn->ssl_)) {
-    int rv;
-
-    if (conn->is_server()) {
-      rv = SSL_accept(conn->ssl_);
-      conn->HandleSSLError("SSL_accept:ClearOut",
-                           rv,
-                           kZeroIsAnError,
-                           kSyscallError);
-    } else {
-      rv = SSL_connect(conn->ssl_);
-      conn->HandleSSLError("SSL_connect:ClearOut",
-                           rv,
-                           kZeroIsAnError,
-                           kSyscallError);
-    }
-
-    if (rv < 0) {
-      return args.GetReturnValue().Set(rv);
-    }
-  }
-
-  int bytes_read = SSL_read(conn->ssl_, buffer_data + off, len);
-  conn->HandleSSLError("SSL_read:ClearOut",
-                       bytes_read,
-                       kZeroIsNotAnError,
-                       kSyscallError);
-  conn->SetShutdownFlags();
-
-  args.GetReturnValue().Set(bytes_read);
-}
-
-
-void Connection::ClearPending(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-  int bytes_pending = BIO_pending(conn->bio_read_);
-  args.GetReturnValue().Set(bytes_pending);
-}
-
-
-void Connection::EncPending(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-  int bytes_pending = BIO_pending(conn->bio_write_);
-  args.GetReturnValue().Set(bytes_pending);
-}
-
-
-void Connection::EncOut(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-  Environment* env = conn->env();
-
-  if (args.Length() < 3) {
-    return env->ThrowTypeError(
-        "Data, offset, and length arguments are mandatory");
-  }
-
-  THROW_AND_RETURN_IF_NOT_BUFFER(args[0], "Data");
-
-  char* buffer_data = Buffer::Data(args[0]);
-  size_t buffer_length = Buffer::Length(args[0]);
-
-  size_t off = args[1]->Int32Value();
-  size_t len = args[2]->Int32Value();
-
-  if (!Buffer::IsWithinBounds(off, len, buffer_length))
-    return env->ThrowRangeError("offset + length > buffer.length");
-
-  int bytes_read = BIO_read(conn->bio_write_, buffer_data + off, len);
-
-  conn->HandleBIOError(conn->bio_write_, "BIO_read:EncOut", bytes_read);
-  conn->SetShutdownFlags();
-
-  args.GetReturnValue().Set(bytes_read);
-}
-
-
-void Connection::ClearIn(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-  Environment* env = conn->env();
-
-  if (args.Length() < 3) {
-    return env->ThrowTypeError(
-        "Data, offset, and length arguments are mandatory");
-  }
-
-  THROW_AND_RETURN_IF_NOT_BUFFER(args[0], "Data");
-
-  char* buffer_data = Buffer::Data(args[0]);
-  size_t buffer_length = Buffer::Length(args[0]);
-
-  size_t off = args[1]->Int32Value();
-  size_t len = args[2]->Int32Value();
-
-  if (!Buffer::IsWithinBounds(off, len, buffer_length))
-    return env->ThrowRangeError("offset + length > buffer.length");
-
-  if (!SSL_is_init_finished(conn->ssl_)) {
-    int rv;
-    if (conn->is_server()) {
-      rv = SSL_accept(conn->ssl_);
-      conn->HandleSSLError("SSL_accept:ClearIn",
-                           rv,
-                           kZeroIsAnError,
-                           kSyscallError);
-    } else {
-      rv = SSL_connect(conn->ssl_);
-      conn->HandleSSLError("SSL_connect:ClearIn",
-                           rv,
-                           kZeroIsAnError,
-                           kSyscallError);
-    }
-
-    if (rv < 0) {
-      return args.GetReturnValue().Set(rv);
-    }
-  }
-
-  int bytes_written = SSL_write(conn->ssl_, buffer_data + off, len);
-
-  conn->HandleSSLError("SSL_write:ClearIn",
-                       bytes_written,
-                       len == 0 ? kZeroIsNotAnError : kZeroIsAnError,
-                       kSyscallError);
-  conn->SetShutdownFlags();
-
-  args.GetReturnValue().Set(bytes_written);
-}
-
-
-void Connection::Start(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-
-  int rv = 0;
-  if (!SSL_is_init_finished(conn->ssl_)) {
-    if (conn->is_server()) {
-      rv = SSL_accept(conn->ssl_);
-      conn->HandleSSLError("SSL_accept:Start",
-                           rv,
-                           kZeroIsAnError,
-                           kSyscallError);
-    } else {
-      rv = SSL_connect(conn->ssl_);
-      conn->HandleSSLError("SSL_connect:Start",
-                           rv,
-                           kZeroIsAnError,
-                           kSyscallError);
-    }
-  }
-  args.GetReturnValue().Set(rv);
-}
-
-
-void Connection::Close(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-
-  if (conn->ssl_ != nullptr) {
-    SSL_free(conn->ssl_);
-    conn->ssl_ = nullptr;
-  }
-}
-
-
-#ifdef SSL_CTRL_SET_TLSEXT_SERVERNAME_CB
-void Connection::GetServername(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-
-  if (conn->is_server() && !conn->servername_.IsEmpty()) {
-    args.GetReturnValue().Set(conn->servername_);
-  } else {
-    args.GetReturnValue().Set(false);
-  }
-}
-
-
-void Connection::SetSNICallback(const FunctionCallbackInfo<Value>& args) {
-  Connection* conn = Unwrap<Connection>(args.Holder());
-  Environment* env = conn->env();
-
-  if (args.Length() < 1 || !args[0]->IsFunction()) {
-    return env->ThrowError("Must give a Function as first argument");
-  }
-
-  Local<Object> obj = Object::New(env->isolate());
-  obj->Set(FIXED_ONE_BYTE_STRING(args.GetIsolate(), "onselect"), args[0]);
-  conn->sniObject_.Reset(args.GetIsolate(), obj);
-}
-#endif
 
 
 void CipherBase::Initialize(Environment* env, Local<Object> target) {
@@ -5803,7 +5094,6 @@ void InitCrypto(Local<Object> target,
 
   Environment* env = Environment::GetCurrent(context);
   SecureContext::Initialize(env, target);
-  Connection::Initialize(env, target);
   CipherBase::Initialize(env, target);
   DiffieHellman::Initialize(env, target);
   ECDH::Initialize(env, target);
