+++ test/parallel/test-tls-external-accessor.js
@@ -16,9 +16,3 @@ const tls = require('tls');
   assert.throws(() => cctx._external, /incompatible receiver/);
   pctx._external;
 }
-{
-  const pctx = tls.createSecurePair().credentials.context;
-  const cctx = Object.create(pctx);
-  assert.throws(() => cctx._external, /incompatible receiver/);
-  pctx._external;
-}
